Tests/LibWeb/Text/expected/reset-text-shadow-after-unhover.txt

hovered text-shadow: rgb(88, 161, 230) 0px 0px 3px
unhovered text-shadow: none
